pyproject.toml
[tool.poetry]
name = "Comments_SPA"
version = "0.1.0"
description = ""
authors = ["Denys Puliaiev <d.s.puliaiev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"
pillow = "^11.0.0"
drf-yasg = "^1.21.8"
whitenoise = "^6.8.2"
django-simple-captcha = "^0.6.0"
channels = "^4.2.0"
bleach = "^6.2.0"
lxml = "^5.3.0"
django-user-agents = "^0.4.0"
daphne = "^4.1.2"
django-redis = "^5.4.0"
channels-redis = "^4.2.1"
python-dotenv = "^1.0.1"
scramp = "^1.4.5"
dj-database-url = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
